Extract workspace render persistence helpers
